requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.3
django-cache-url==3.4.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
environs==11.0.0
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.21.1
openai==1.14.1
packaging==23.2
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
